pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "chaostoolkit-consul"
dynamic = ["version"]
description = "Template project for Chaos Toolkit extensions"
authors = [
    {name = "Chris Doyle", email = "uk.chris.doyle+cdsre@gmail.com"},
]
dependencies = [
    "chaostoolkit-lib>=1.37",
    "py-consul>=1.4.0",
]
requires-python = ">=3.7"
readme = "README.md"
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: Freely Distributable",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation",
    "Programming Language :: Python :: Implementation :: CPython"
]

[project.urls]
documentation = "https://chaostoolkit.org"
repository = "https://github.com/cdsre/chaostoolkit-consul"
changelog = "https://github.com/cdsre/chaostoolkit-consul/blob/master/CHANGELOG.md"

[tool]

[tool.pdm]
version = { source = "scm" }

[tool.pdm.scripts]
lint = {composite = ["ruff chaosconsul/", "isort --check-only --profile black chaosconsul/", "black --check --diff chaosconsul/", "mypy chaosconsul/", "bandit -r chaosconsul"]}
format = {composite = ["isort --profile black chaosconsul/", "black chaosconsul/", "ruff chaosconsul/ --fix"]}
test = {cmd = "pytest"}

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.0.280",
    "black>=23.3.0",
    "pytest-sugar>=0.9.7",
    "pytest-cov>=4.1.0",
    "pytest>=7.4.0",
    "coverage>=7.2.7",
    "isort>=5.11.5",
    "mypy>=1.4.1",
    "bandit>=1.7.5",
]

[tool.black]
color = true
line-length = 80
target-version = ['py310']

[tool.mypy]
namespace_packages = true

[tool.ruff]
line-length = 80
target-version = "py310"

[tool.isort]
line_length = 80
known_first_party = 'chaosconsul'
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
addopts = "-v -rxs --cov chaosconsul --cov-report term-missing:skip-covered  --cov-report xml -p no:warnings"
